pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cookiesncream"
version = "0.0.1-beta"
authors = [{ name = "Benjamin Liden", email = "lidenb@oregonstate.edu" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">3.10"
dependencies = [
    "pillow",
    "pillow_lut >= 1",
    "pydantic >= 2",
    "pydantic-settings >= 2",
    "zmq",
]

[project.urls]
Homepage = "https://github.com/bcliden/cookiesncream"
Issues = "https://github.com/bcliden/cookiesncream/issues"

[tool.setuptools.package-data]
"cookiesncream.lut" = ["*.cube"]

[project.scripts]
"cookiesncream.zmq" = "cookiesncream.zero_mq.entry:main"
"cookiesncream.cli" = "cookiesncream.cli.entry:main"

[project.gui-scripts]
# do we have any of these?

[[tool.mypy.overrides]]
module = "pillow_lut"
ignore_missing_imports = true
